Add tzinfo-data and fix tests
The previous attempt to use stub_const did not "unset" between runs.
It is simpler to hard-code a timezone for all tests, to guarantee
that there are no ambiguities.

Because SQL timestamps do not allow for timezone or offset data, this
should not be a limiting factor for the tests.