Add options in the config to control how the character encoding will work.
Added a param named enable_encoding, false by default that will be used
to enable/disable character conversion to UTF-8.

Added a list where we can specify the columns we want to convert or
empty if we aim to convert all (empty by default). Adding a column to
this list will only convert the selected column.

The converion method is selected according to the encoding the string is
coming from the database.

Addresses concerns of #143 (backport of #151)

Fixes #152