Add bool config for tracking column as timestamp
This allows plugin users to set `use_column_value` on tracking columns
where the column type is a timestamp. Currently, this is only possible
using a numeric column, but there are use cases where the user wants to
select rows from a table where the column value is greater than the last
seen timestamp value of that column.

It turns out that this behaviour works fine with simply setting
`use_column_value` to true on the required column. The only change
needed is to add a flag indicating that on a clean run, the default min
value should be set to `Time.at(0).utc` rather than `0`.

This should close logstash-plugins/logstash-input-jdbc#125