Convert sql_last_value from DateTime to Time
There's an ordering issue with the tests. For some reason, when `when
fetching time data with jdbc_default_timezone set` is run before the new
test `when iteratively running plugin#run with timestamp tracking column
with column value`, the new test fails. This turns out to be because the
`jdbc_default_timezone` setting causes Sequel to return `DateTime`
objects rather than `Time` objects (default). I'm not really sure why
this behaviour leaks out of the first test, but the failure only occurs
if they're run in this order.

To fix, this commit just checks to see if `tracking_column_type` is
"timestamp", `use_column_value` is true and that the value returned from
Sequel is a `DateTime`. If so, it converts to a `Time`.