Add bool config for tracking column as timestamp (#156)
* Add bool config for tracking column as timestamp

This allows plugin users to set `use_column_value` on tracking columns
where the column type is a timestamp. Currently, this is only possible
using a numeric column, but there are use cases where the user wants to
select rows from a table where the column value is greater than the last
seen timestamp value of that column.

It turns out that this behaviour works fine with simply setting
`use_column_value` to true on the required column. The only change
needed is to add a flag indicating that on a clean run, the default min
value should be set to `Time.at(0).utc` rather than `0`.

This should close logstash-plugins/logstash-input-jdbc#125

* Switch to more flexible tracking_column_type

* Convert sql_last_value from DateTime to Time

There's an ordering issue with the tests. For some reason, when `when
fetching time data with jdbc_default_timezone set` is run before the new
test `when iteratively running plugin#run with timestamp tracking column
with column value`, the new test fails. This turns out to be because the
`jdbc_default_timezone` setting causes Sequel to return `DateTime`
objects rather than `Time` objects (default). I'm not really sure why
this behaviour leaks out of the first test, but the failure only occurs
if they're run in this order.

To fix, this commit just checks to see if `tracking_column_type` is
"timestamp", `use_column_value` is true and that the value returned from
Sequel is a `DateTime`. If so, it converts to a `Time`.