connect our logger up to the charset converters
The `LogStash::Util::Charset` utility will attempt to send a `warn` to its
`@logger` instance variable when it encounters invalid encoding, but does not
take a logger to its constructor.

Failing to wire up a logger means that each time we attempt to convert a string
and fail to arrive at valid UTF-8, we send `warn` to `NilClass`, which raises
a `NoMethodError`.